+++ src/components/atoms/TimeDurationInput/TimeDurationInputBase.tsx
@@ -2,7 +2,7 @@ import {
   type Duration,
   parseISO8601Duration,
   toISO8601Duration,
-} from '@/isoDuration';
+} from '@/utilities/isoDuration';
 import {
   Box,
   FormControl,
